Move blanket Handler impl into grpc.rs
